faire du front maintenant

+++ db/migrate/20201029141158_create_quantities.rb
@@ -1,7 +1,7 @@
 class CreateQuantities < ActiveRecord::Migration[6.0]
   def change
     create_table :quantities do |t|
-      t.integer :number
+      t.float :number
       t.references :recipe, null: false, foreign_key: true
       t.references :ingredient, null: false, foreign_key: true
